+++ crates/entrypoint/src/cli/command/azure_devops/agent/pool/mod.rs
@@ -2,3 +2,4 @@ mod azure_devops_agent_pool_cli;
 pub use azure_devops_agent_pool_cli::*;
 pub mod entitlement;
 pub mod list;
+pub mod summary;