+++ crates/entrypoint/src/cli/command/azure_devops/agent/pool/summary/mod.rs
@@ -1,0 +1,2 @@
+mod azure_devops_agent_pool_summary_cli;
+pub use azure_devops_agent_pool_summary_cli::*;